pyproject.toml
[project]
name = "info698-capstone"
version = "0.1.0"
description = "Agent Based RAG system over PMC"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.45",
    "langchain-openai>=0.3.8",
    "langchain-weaviate>=0.0.4",
    "lxml>=5.3.1",
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=4.0.1",
    "streamlit>=1.42.2",
    "tabulate>=0.9.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "watchdog>=6.0.0",
    "weaviate-client>=4.11.3",
]
